Break api client into smaller pieces
