+++ src/designer/elsa-workflows-designer/src/index.ts
@@ -2,6 +2,6 @@ import 'reflect-metadata';
 
 export {Components, JSX} from './components';
 export {Container} from 'typedi';
-export {ElsaApiClientProvider, createElsaClient} from './services/elsa-api-client';
+export {ElsaApiClientProvider, createElsaClient} from './services/api-client/api-client';
 export {stripActivityNameSpace} from './utils';
 export {PluginRegistry, ActivityNameFormatter} from './services';
