+++ src/designer/elsa-workflows-designer/src/services/api-client/activity-descriptors-api.ts
@@ -1,0 +1,23 @@
+import {ActivityDescriptor} from "../../models";
+import {AxiosInstance} from "axios";
+
+export interface ActivityDescriptorsApi {
+  list(): Promise<Array<ActivityDescriptor>>;
+}
+
+export interface ActivityDescriptorResponse {
+  activityDescriptors: Array<ActivityDescriptor>;
+}
+
+export class ActivityDescriptorsApiImpl implements ActivityDescriptorsApi {
+  private httpClient: AxiosInstance;
+
+  constructor(httpClient: AxiosInstance) {
+    this.httpClient = httpClient;
+  }
+
+  async list(): Promise<Array<ActivityDescriptor>> {
+    const response = await this.httpClient.get<ActivityDescriptorResponse>('descriptors/activities');
+    return response.data.activityDescriptors;
+  }
+}
