+++ src/designer/elsa-workflows-designer/src/services/api-client/descriptors-api.ts
@@ -1,0 +1,17 @@
+import {ActivityDescriptorsApi, ActivityDescriptorsApiImpl} from "./activity-descriptors-api";
+import {AxiosInstance} from "axios";
+
+export interface DescriptorsApi {
+  activities: ActivityDescriptorsApi;
+}
+
+export class DescriptorsApiImpl implements DescriptorsApi {
+  httpClient: AxiosInstance;
+  activities: ActivityDescriptorsApi;
+
+  constructor(httpClient: AxiosInstance) {
+    this.httpClient = httpClient;
+    this.activities = new ActivityDescriptorsApiImpl(httpClient);
+  }
+
+}
