+++ src/designer/elsa-workflows-designer/src/services/api-client/labels-api.ts
@@ -1,0 +1,13 @@
+import {AxiosInstance} from "axios";
+
+export interface LabelsApi {
+
+}
+
+export class LabelsApiImpl implements LabelsApi {
+  private httpClient: AxiosInstance;
+
+  constructor(httpClient: AxiosInstance) {
+    this.httpClient = httpClient;
+  }
+}
